# apps/docs-site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Claustrum Memory Core API",
    "version": "0.1.3",
    "description": "Auto-generated from memory-core route declarations and inline Zod validations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/v1/api-keys": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "GET /v1/api-keys",
        "operationId": "get_v1_api_keys",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/api-keys-routes.ts"
      },
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "POST /v1/api-keys",
        "operationId": "post_v1_api_keys",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/api-keys-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "workspace_key": {
                    "type": "string"
                  },
                  "device_label": {
                    "type": "string"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "device_label"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/api-keys/{id}/revoke": {
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "POST /v1/api-keys/:id/revoke",
        "operationId": "post_v1_api_keys_id_revoke",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/api-keys-routes.ts"
      }
    },
    "/v1/api-keys/one-time/{token}": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "GET /v1/api-keys/one-time/:token",
        "operationId": "get_v1_api_keys_one_time_token",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "x-source-file": "apps/memory-core/src/http/routes/api-keys-routes.ts"
      }
    },
...
    "/v1/users/{userId}/api-keys": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "GET /v1/users/:userId/api-keys",
        "operationId": "get_v1_users_userId_api_keys",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/api-keys-routes.ts"
      }
    },
    "/v1/users/{userId}/api-keys/reset": {
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "POST /v1/users/:userId/api-keys/reset",
        "operationId": "post_v1_users_userId_api_keys_reset",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/api-keys-routes.ts",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
...
}